+++ CHANGELOG.md
@@ -4,7 +4,14 @@ All Notable changes to `mapudo/guzzle-bundle` will be documented in this file.
 
 Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
-<!-- @TODO - Change date accordingly -->
+Note: This is WIP and will be edited when the other issues for the 1.1.0 milestone are finished as well.
+
+## [1.1.0] - 2017-03-10
+### Changed
+- Requests that are logged with the LogMiddleware, will now be logged with a client name
+- Symfony Profiler Handler now allows filtering by client
+- Symfony Profiler Handler now shows the client as a label on the request
+
 ## 2017-03-03 
 
 ### Added